refactored some components to make them smaller. I may even try a MERGE
