+++ TradesMin_App/src/Components/ProjectsPagesV2/EditButtons/EditButton.jsx
@@ -1,14 +1,12 @@
-import { useNavigate } from 'react-router-dom';
-
-
+import { useNavigate } from "react-router-dom";
 
 const EditButton = ({ project }) => {
-const navigate = useNavigate()
-console.log("project to EDIT:", project)
+  const navigate = useNavigate();
+  console.log("project to EDIT:", project);
 
-    const handleEditClick = () => {
-        navigate(`/editproject/${project.id}`, { state: { project } });
-    };
+  const handleEditClick = () => {
+    navigate(`/editproject/${project.id}`, { state: { project } });
+  };
   return (
     <button
       onClick={handleEditClick}
